64 bit timestamp

+++ src/can_common.h
@@ -140,7 +140,7 @@ public:
     BytesUnion data;    // 64 bits - lots of ways to access it.
     uint32_t id;        // 29 bit if ide set, 11 bit otherwise
     uint32_t fid;       // family ID - used internally to library
-    uint32_t timestamp; // CAN timer value when mailbox message was received.
+    uint64_t timestamp; // CAN timer value when mailbox message was received.
     uint8_t rtr;        // Remote Transmission Request (1 = RTR, 0 = data frame)
     uint8_t priority;   // Priority but only important for TX frames and then only for special uses (0-31)
     uint8_t extended;   // Extended ID flag
@@ -156,7 +156,7 @@ public:
     BytesUnion_FD data;   // 64 bytes - lots of ways to access it.
     uint32_t id;          // EID if ide set, SID otherwise
     uint32_t fid;         // family ID
-    uint32_t timestamp;   // CAN timer value when mailbox message was received.
+    uint64_t timestamp;   // CAN timer value when mailbox message was received.
     uint8_t rrs;          // RRS for CAN-FD (optional 12th standard ID bit)
     uint8_t priority;     // Priority but only important for TX frames and then only for special uses. (0-31)
     uint8_t extended;     // Extended ID flag
